+++ src/App.jsx
@@ -4,7 +4,6 @@ import { QueryClient, QueryClientProvider } from 'react-query';
 import Header from './components/Header';
 import Footer from './components/Footer';
 import Nav from './components/Nav';
-// import Home from './pages/Home/Home';
 import Restaurant from './pages/Restaurant/Restaurant';
 import HomePage from './pages/HomePage/HomePage';
 import Play from './pages/Play/Play';
@@ -12,7 +11,9 @@ import Detail from './pages/Detail/Detail';
 import Login from './pages/Login/Login';
 import Join from './pages/Join/Join';
 import MyPage from './pages/MyPage/MyPage';
-import Judge from './pages/Judge/Judge';
+import JudgeHome from './pages/Judge/JudgeHome';
+import JudgeNew from './pages/Judge/JudgeNew/JudgeNew';
+import JudgeList from './pages/Judge/JudgeList/JudgeList';
 
 const queryClient = new QueryClient();
 
@@ -25,7 +26,9 @@ const App = () => {
         <Route path="/" element={<HomePage />} />
         <Route path="/restaurants" element={<Restaurant />} />
         <Route path="/play" element={<Play />} />
-        <Route path="/restaurants-judge/*" element={<Judge />} />
+        <Route path="/restaurants-judge" element={<JudgeHome />} />
+        <Route path="restaurants-judge/new-judge" element={<JudgeNew />} />
+        <Route path="restaurants-judge/list" element={<JudgeList />} />
         <Route path="/login" element={<Login />} />
         <Route path="/join" element={<Join />} />
         <Route path="/mypage" element={<MyPage />} />
