+++ src/pages/Judge/JudgeNew/JudgeNewOutModal.jsx
@@ -2,9 +2,9 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import { reset } from '../../redux/judgeNew';
-import { OutModal, OutModalInner } from './judge.style';
-import xIcon from '../../assets/img/x-icon.svg';
+import { reset } from '../../../redux/judgeNew';
+import { OutModal, OutModalInner } from '../judge.style';
+import xIcon from '../../../assets/img/x-icon.svg';
 
 const JudgeNewOutModal = ({ setIsOut }) => {
   const navigate = useNavigate();
